+++ packages/flame/CHANGELOG.md
@@ -10,7 +10,9 @@
  - Add `SpriteButton.asset`
  - Add `NineTileBox.asset`
  - Add `BaseGame.remove` and `BaseGame.removeAll` helpers for removing components
+ - Add `BaseComponent.remove` and `BaseComponent.removeAll` helpers for removing children
  - Rename `Camera.cameraSpeed` to `Camera.speed`
+
 
 ## [1.0.0-releasecandidate.13]
  - Fix camera not ending up in the correct position on long jumps